Updated POST handling to read using InputStream.
Grizzly is exhibiting weird behaviours when processing chunked POST
requests with 100-Continue.  It leaves the first chunk byte size in the
input stream which breaks automatic entity processing.  I was forced to
write code to parse the initial stream until the start of XML data as
marked by "<?".  When this is fixed we can return to full entity
processing.

Now with the Apache HTTP client using chunked POST requests with
100-Continue we can avoid the Apache httpd post limitation when
renegotiating the SSL connection.